Add infrastructure to check for problems in Markdown files
Tasks are provided to check for problems in the project's Markdown files:

- Use markdownlint to check for common problems and formatting.
- Use markdown-link-check to check for broken links.

A GitHub Actions workflow is included to automatically run the tasks. The workflow is triggered on any push or pull that
changes relevant project files, in order to avoid the introduction of problems with the project filesystem. It is also
triggered periodically, in order to catch breakage caused by external changes.